Adds a worker pool (#2)
In this PR, we add a worker pool to ABQ. The idea (for now) is that

we have a number of worker methods, for example to shell out to rspec and run tests. From the point of view of ABQ, these can be thought of as pure functions (they don't affect ABQ's state, they just return something that ABQ then acts on). Their call signature, found in workers/protocol/src/lib.rs, is

```
pub trait Worker {
    type Input;

    fn run(input: Self::Input) -> Output;
}
```

for demonstration here, I've added only the Echo worker, which behaves like bash echo
we may want to change this model in the future, for example to have stronger coupling to the queue as discussed in slack. For now though, I think this is a good starting point.
these worker methods are multiplexed on based on the enum WorkAction, which corresponds to a unit of work. Right now the only WorkAction is Echo, but in the future we'll include Rspec, Jest, etc. Maybe a plugin model is a good idea here?

to act on units of work popped from the queue, and execute them in these worker methods, we now have a WorkerPool. This a pool of N threads and processes, where N is the number of workers we want. A request to execute a unit of work is sent to the worker pool; the first thread to lock and take the message gets to process it.

Each worker thread figures out the worker method that needs to be run, and executes it in the process pool. Once a result is received (or the process times out, or the process fails too many times), the worker thread sends the result back to the queue across a unix socket, and continues to poll for new work.

The reason for a process pool here is mostly to isolate the worker processes from the main abq/worker pool process, since e.g. we can't really anticipate what all might happen when we run tests in an arbitrary framework.