Support invoking and retrieving results from an ephemeral ABQ instance (#3)
* Simple cli for echo commands

* Use invocation mechanism to manage multiple invocations of abq

With this commit, we add an idea of "invokers" of abq, which send a set
of work units to the queue server and wait for their results. Now, a
unit of work in the queue is a triple of (invocation ID, work ID, action).
The invocation ID is abq-internal and used only for coordination of work
that should be sent back to particular invokers. The work ID is opaque
to abq and used only by invokers to line up sent units of work to their
results.

This lays the groundwork for sending work to an ephemeral queue from
external processes.

* Make ABQ take a socket path rather than always creating a temp socket

* Add ephemeral and temporary queue mode to echo examples

* Pretty-print output of abq-invoked work

* Clippy

* Avoid unneeded allocations

* Simple "Exec" worker spawns a process and returns its status

* Adds a cargo test runner

This commit introduces a simple implementation of Rust's `cargo` test
runner with `abq`. `abq cargo test` discovers all Rust tests in the
directory `abq`  was invoked from, and then runs each tests in a
separate `cargo` subprocess using the abq queue/worker pool. Results are
returned back to the invoking process with a simple OK/FAIL message.

Of course there are many things to improve on here, but hopefully it's a
good demonstration of what we can do and plug in now. A screencapture
attached to the PR for this commit demonstrates the behavior.

* Update lockfile

* Don't use exec worker in test imports

* Remove redundant clones

* Update bin/src/main.rs

* Update lib/src/invoke.rs

* Split up echo, cargo into plugins and move `run_work` to its own module