Adds TLS support (#111)
* Implement TLS support for ABQ

This patch adds new TLS-backed implementations of the four
client/server interfaces now used by ABQ entities.

For simplicity in our early days, I've checked in a cert and private key
that all TLS-based servers will use, and all TLS-based client will
authenticate. Of course, in a real production usage we'd want to
configure that out-of-band and use different certs/server names for the
various servers, and actually have private keys be private.

This patch adds two dependencies.

First first is `rustls`. This is the TLS implementation in rust that has won.
I did some basic auditing and am confident in the quality of the dependency.

- The maintainers had an audit done by Cure53 in 2020: https://github.com/rustls/rustls/blob/master/audit/TLS-01-report.pdf.
  The scope of the audit included the TLS library and its two main
  dependencies, `ring` (for cryptography) and `webpki` (for cert
  verification). The results of the audit were positive.
- There don't appear to be any significant issues or vulnerabilities
  currently known and unpatched in any of the libraries.
- We have three licensing options: MIT, Apache 2, or ISC

It's possible we may want to go with another underlying TLS
implementation (e.g. use bindings to BoringSSL), but at least right now,
`rustls` has been rather straightforward to use, and appears to work
correctly. So it's been nice for our prototyping.

The second dependency is `tokio-rustls`. This is a much more immature
dependency that provider plumbing between the TLS+TCP abstractions `rustls`
provides and the `tokio` async runtime that we've been using for some
things in ABQ. I've vendored the dependency because I would prefer that
we do not maintain our own networking + runtime plumbing for the
application layer of ABQ.

I've read the source code and it seems reasonable; it is quite a small
library, so if we run into problems, we can either fork the library or
maintain our plumbing so that TLS and the async runtime play nice. For
now I think it's more productive for us to use this library; ABQ's
networking is already non-trivial, and it seems to hold up well.

I briefly surveyed what other Rust-based projects with networking do;
there aren't a ton of samples, but SourceGraph seems to suggest that
folks either maintain this themselves, or use a dependency like
`tokio-rustls`.

Finally, the maintainers of `tokio-rustls` are both well-known and
well-regarded; one maintainer is the same one who made `rustls` and its
underlying primitives. The bug tracker suggests the project is in active
development and support. Our licensing option is MIT.

* Print encryption: TLS in abq version, when built with TLS

* Set up CI for testing the TLS implementation

* Clippy

* Add note making new cert/key pair

* Don't test TLS in the non-tls CI run

* Pull the native test runners before testing tls

* Don't write logs to stderr when ABQ_LOG is unset

* Dogfood and send abq test results for tls tests

* Check in tls dogfood make command

* Update job-name for Captain