Add an auth token header strategy (#114)
This patch adds an token-based authentication strategy for abq clients
and servers. As discussed elsewhere, right now I've just implemented a
constant token for servers to use in authenticating/authorizing all
clients - there is no concept of roles or anything like that.

In the next patches, I intend to make the authentication token
configurable at the time a queue, workers, and `abq test` is
instantiated, via a CLI flag (we can add environment variables and so
forth later on as well, if needed). Callers of `abq test` and `abq work`
will have to know what token, if any, the queue expects; for now, I
think we can hard-code this as a secret in our CI set-up and development
environments.

Okay, so about the auth token. I wanted to keep this really simple; as I
mentioned what we currently do is nothing more than check whether a
client passes, in the header of a connection, the exact token a server
is configured with. To build a base that we can build off of, the token
is a cryptographic psuedo-randomly generated buffer of 30 bytes sampled
from the ascii alphanumeric alphabet. That's the size of token GitHub
uses and it yields 178 bits of entropy, which I think is far more than
enough for us (especially right now). We seed the RNG with the system
cryptographic RNG (i.e. `unrandom` on unix).

I do think all of that is overkill right now, but I think it's in line
with what we'll want for auth tokens longer-term.