Allow configuring auth token for abq runs from the command line (#115)
* Provide a way to get an auth token from the CLI

* Add tests among queue servers regarding auth'd connections

* Allow configuring auth token for abq runs from the command line

This patch now enables us to pass an auth token when you start a queue
with `abq start`, to workers with `abq work`, and to `abq test`. To get
a fresh token for these uses, a new subcommand `abq token new` has been
added.

With this, we now also test that authorization with the tokens works as
expected. There are some unit tests that expand the coverage of the
tests from the parent of this commit.

More importantly, I've made it so that our integration tests run over all
combination of authentication options abq provides.
Right now that's just whether there's an auth token or not; in the future
it'll also be whether TLS should be enabled.

The way the new integration tests work is we have a macro that runs a
concrete tests over all combination of network options. Now, the
implementation of that isn't the prettiest thing in the world... and it
might be quite hard to understand, please let me know if it is! I
normally wouldn't do such a thing, but I think we get a lot of milage
out of multiplexing over tests in this way.