Move TLS handshakes out of the hot path (#135)
With this patch we move TLS handshakes and authentication out of the
hot-path of accepting connections to the servers, which must always
happen on the main threads. Instead we defer handshake and auth to
actually handling the accepted connection. To do this we now say that
accepting a connection yields an opaque "unverified stream". You can't
do anything with an unverified stream other than to make it into a
verified stream, which you do by calling a server's "handshake". In
practice that method has its functionality on the tin; we do a TLS
handshake where appropriate, or no handshake if you're running a raw TCP
queue.

Closes #127