Have workers exit with non-zero code if any test in a run failed (#137)
* Workers exit with non-zero code if any test in a run failed

To support use cases where you'd like to re-run all processes involved
in an ABQ run if any test in that run fails (e.g. in CI), this patch now
makes it so that workers poll the queue for the final result of the run
before exiting. They then use that result to determine whether to exit
cleanly or with a failing code.

Workers will poll the queue every second for the run result, for now. We
can change that polling frequency if the future if we'd like.

Closes #123

* Add integration test for workers exiting loudly on any run's test failure

* Read all output from abq-cargo as piped

I don't know why, but for some reason before this, the process would
hang!

* Run CLI integration tests in serial

I am afraid there is a race here somewhere, and also, running them in
parallel eats CPU because many child processes are started.

* Allow redundant closure calls in macros

* Drop stream rather than explicitly shutting it down

The client stream may be dead by the time we call shutdown, so don't try
to shut down explicitly, just drop it.

* Update difference after merge with main