Resolve some failure conditions (#153)
* Do not block the work server when manifest is not yet ready

When a worker asks the work server for the next test to run, rather than
the work server keeping the connection open until the manifest is
available, immediately respond with either (1) the next set of tests to
run, or (2) a message that the manifest is still pending. In the latter
case, it's now the responsibility of the worker to re-query the server
rather. This patch implements subsequent queries with an exponential
decay.

* Stream results to a client only after a client is definitely listening

This avoids a race condition wherein we do not properly associate how to
send test results to an `abq test` client before the first test result
is attempted to be sent back.

Closes https://github.com/rwx-research/abq/issues/152

* Rename reconnect_with to update_connection_to

* Add back unintended omissions