Improve latency of test result drainage (#154)
* Take unique locks on per-run-id responders rather than the map of responders

The `abq` main server maintains a map of run id -> `abq test`
connections that is used to determine how test results should be
reported back to the client. We must keep a lock over this map for two
reasons:

1. When new test runs are invoked, the map needs to be updated; when we
   complete a test run, the map also needs to be updated to remove
   associations
2. When a test result is being sent back, we must make sure two tests
   are not written at the same time to the same stream

Previously we resolved both constraints by keeping a single mutex over
the map. But this is inefficient, because in the case where we have
multiple parallel runs, we want uncontested access to send results to
back to both test A and B, since sending a test result does not require
modification of the associations. So now, this patch keeps the
associations behind a read-write lock (making the non-modification path
uncontested for result reporting), and a mutex over the underlying
connection.

The results are pretty good, if I may say. With three workers on my
machine, before this patch:

```
captain:
  average: 111.66687838236491
  max: 112.10106921195984
  min: 111.42113709449768
```

After this patch:

```
captain:
  average: 58.55347498257955
  max: 59.123817920684814
  min: 58.09637498855591
```

So a -50 second improvement, on average. Now, there is contention
between the workers running on my machine - but that is not the
important part of this statistic. Before and after the workers take
about the same time to complete, and what remains, is the delta in how
long it takes us to drain test results back to `abq test`.

Part of #128

* fixup! Take unique locks on per-run-id responders rather than the map of responders

* fixup! Take unique locks on per-run-id responders rather than the map of responders

* Buffer test results sent back to `abq` clients

Rather than sending back test results one-at-a-time, we should batch
them. This will improve latency of abq test runs, and not decrease the
interactivity of `abq test` too much if we keep the buffer sizes
relatively small compared to the size of an entire test suite.

Moreover, in CI use cases, you probably don't care too much about
interactivity of test results - you would rather get the test results
ASAP, so the batch sizes can be increased in those contexts.

For now, this patch batches test results 10 at a time.
In future patches we can improve how we decide what the batching size
should be.

In local testing of Captain's test suite running on abq, this has provided
a -10 second improvement in the wall time compared to e9ca2fea3dc6ea1a83e69f120262057fc5bfde27.
Now, when running `abq` in debug mode, `abq test` finishes around the
same time that the workers do, rather than taking much longer.