Graceful shutdown handling (#211)
* Rename to immediate shutdown

* Abstract away shutting down servers behind a ShutdownManager

* Add a retirement mode where in queue rejects new requests for work to do

This will be used in hosted provisioning, where we'd like to drain a
queue before shutting it down. As part of this, the queue should stop
accepting new requests; in this patch, it logs warning and refuses
connections if new test run requests come in after retirement.

Note that retirement is sequentially consistent but is not atomically
consistent. This means that it is possible to observe a request flow
like

```mermaid
sequenceDiagram
    participant Admin
    participant Supervisor
    participant Queue

    rect rgb(191, 223, 255)
    Admin->>+Queue: snd Retire
    rect rgb(200, 150, 255)
    Supervisor->>+Queue: snd NewTest
    Queue-->>-Supervisor: ack NewTest
    end
    Queue-->>-Admin: ack Retire
    end
```

or

```mermaid
sequenceDiagram
    participant Admin
    participant Supervisor
    participant QueueHdl1
    participant QueueHdl2

    Admin->>+QueueHdl1: snd Retire
    Supervisor->>+QueueHdl2: snd NewTest
    QueueHdl1->>QueueHdl1: record retirement
    QueueHdl1-->>-Admin: ack Retire
    QueueHdl2->>QueueHdl2: record NewTest
    QueueHdl2-->>-Supervisor: ack NewTest
```

However, this both of these cases are okay for us for now, because
retirement does not mean immediate shutdown - rather it just means we
can start polling for when the queue is completely drained.

A follow-up will add retirement via SIGQUIT as well.

Closes #200

* s/new/default