Have queue go into retirement on SIGTERM/SIGINT/SIGQUIT (#212)
* Gracefully shutdown, then terminate queue on SIGTERM-like signals

Closes #200

* Add some debug logging during signal shutdown

* Have an external queue admin determine when the queue should die

The queue shouldn't exit just because it's drained, because there might
be additional workers it doesn't know about looking to connect after a
queue is drained but before it's exited. Let's just have the queue
admininstrator decide when the queue should be killed instead. Receiving
two termination signals will always kill the queue.

* Address lints