Implement a better lock semantics for thread-shared queues (#236)
* Upgrade to rust 1.65

* Rename queues to run_state and update comments

* Implement a better lock semantics for thread-shared queues

This patch prepares for #185.

Previously we represented queues to run IDs via the following data
structure

```
Mutex<
  Map<RunId, RunState>,
  Map<RunId, RunData>,
>
```

`RunState` represents the current state of a given test run, which can
roughly be:

- The run has yet to start
- The run is waiting for a manifest
- The run is active, and the next test to run in the manifest is pointed
  to
- The run has completed or was cancelled

`RunData` is additional metadata for a test run that we keep around only
while a test run is active, since we no longer need it after run
completion. When a run is done we drop the additional data to avoid
memory leaks.

This representation has been okay for us so far, because indexes into
the queues are write-heavy, but it can be better. For one, there is no
need to take a lock over all queues for all test runs when a worker just
wants to pull tests for one test run ID. For another, this model will be
less useful once we add some level of support for
https://github.com/rwx-research/abq/issues/185, since that will require
us to read back into the `RunQueues` far more often than we currently do.

With that context, this PR implements a new model for the shared queues:

```
RwLock<
  Map<RunId, Mutex<{
    state: RunState,
    data: Optional<RunData>,
  }>>
>
```

In this model you now only need to take a read guard to index into
queues; since writes to the containing queues data structure are
relatively rare (they happen only when you launch a new test run), this
makes operating over indepdent test runs suffer less contention.
You now only need to lock to access the state and data for an individual
test run. (With #185 the nested mutex will also likely become a
read-write lock).

With this change we also decrease the stack size of `RunData` to avoid
empty representations taking up more memory than they need to.

I haven't actually measured the perf characteristics of this, but it
should be strictly positive. We can run some measurements again after we
have hosted infra set up, but as a positive signal, we did a similar
transformation to the representation of supervisors on the queue side
about a month ago, and got substantial wins in the perf characteristics
of the queue.

* Store estimate test runs on toplevel

* Revert native runner changes

* Address lint

* Bump timeout

Since we're updating the rust version, the cache is dead and will take
longer.