`RunTimeoutManager` keeps track of test runs registered for timeout (#243)
* `RunTimeoutManager` keeps track of test runs registered for timeout

The `RunTimeoutManager` data structure is responsible for managing a
bucket of timeouts, and exposing a convenience API to fire when any
timeout fires. This will be used by the queue to introduce timeouts for
tests that have ended but have not yet had all of their test results
reported.

Insertion of timeouts is relatively cheap and contests only with tasks
that are attempting to wait for timeouts, which in practice will be one
main task in the queue. Waiting for a timeout requires exclusive access,
but yields at regular intervals (currently 10 microseconds).

In support of #185

* Allow dead code for now