Cancels of test runs if final results don't come in 2 minutes (#244)
* Cancels of test runs if final results don't come in 2 minutes

This is the first pass of what's suggested and discussed in #185. After
we hand out the last test for a test run to a worker, we start a 2
minute timer (that can be configured later, maybe coming from the
supervisor options). The timer will fire unconditionally; if it fires
while the test run is still active, we notify the supervisor that the
test should be failed due to a timeout. If the test is already done or
cancelled for another reason, the timeout is dropped with no
consequence.

Care is taken to make sure that we don't interfere with manual user
cancellations or completed test results that race with the timeout
notification.

Addresses #185

* Address lints

* Add test for cancelled test runs after timeout

* Add more logs for final test results and timeouts

* Span trace run completions