Persist NextTest request connections between workers and queue  (#339)
* Queue implements persistent connections for worker tests requests

* Rename GetNextWorkBundle

* Persist NextTest request connections between workers and queue

This patch implements persistent connection holding between the worker
server and target workers. If a worker loses a connection to the work
server, it will retry the connection up to three times.

There is a bit more locking overhead in here than we really need, but
that can be addressed in the future. In any case it should bring us
closer to the goals described in #281.

Part of #281

* Add some comments