Enforce that a type-tag is forced to be present when deserializing incremental results (#414)
* Enforce that a type-tag is forced to be present when deserializing incremental results

Previously, we expressed incremental test result messages as separate
tagged structs joined as an untagged union. In particular, the "done"
message for incremental results has shape

```
{ type: 'incremental_result_done', last_test_result: Option<TestResult> }
```

and had serde model definition

```
 #[derive(Serialize, Deserialize)]
 #[serde(tag = "type", rename = "incremental_result_done")]
pub struct IncrementalTestResultDone {
    pub last_test_result: Option<TestResult>,
}
```

It appears that serde does not enforce the tag to be present when
deserializing a struct! This means that the following was admitted
currently:

```
{}
```

or in fact, all json objects, since `type` did not need to be present,
and `last_test_result` is optional. [See this playground](https://play.rust-lang.org/?version=stable&mode=debug&edition=2021&gist=2f24c20d5ba6fe61fd6842ef4d003295).

The solution is to tag on the incremental test result enum, that is,
form

```
 #[serde(tag = "type")]
pub enum IncrementalTestResultMessage {
    #[serde(rename = "incremental_result")]
    One { one_test_result: TestResult },
    #[serde(rename = "incremental_result_done")]
    Done {
        last_test_result: Option<TestResult>,
    },
}
```

In this model, serde will indeed enforce the tag to be present.
[See this playground](https://play.rust-lang.org/?version=stable&mode=debug&edition=2021&gist=2f24c20d5ba6fe61fd6842ef4d003295).

TEST:
  - Add tests on the CLI test side to ensure we are catching the number
    of run tests
  - Add unit tests to make sure only the messages we want pass through

* Make sure v0_2 is parsed first

* Reflect pass/fail status in tests

* Add unit tests for protocol parsing

* Apply lints