Implement a cancel-safe version of reading from a stream asynchronously (#427)
* Implement a cancel-safe version of reading from a stream asynchronously

While working on https://github.com/rwx-research/abq/pull/426, I
came across a usage pattern that introduced that we haven't yet handled
before but will need to in order to support that PR. The problem has to
do with cases where we opportunistically branch on the first async task
to resolve out of multiple; in particular, in #426, we have a pattern
like

```
loop {
    select! {
        read_message_from_queue() => { .. }
        wait_500_ms() => { .. }
    }
}
```

where if we receive a message from the queue first, we handle that; but
if we sit around for 500ms, we go do something else, and then come back
and try to read the next message.

The issue here is when the `wait_500_ms` branch completes while
`read_message_from_queue` is executing. Currently, our read function
`async_read` is not "cancel-safe" - when `wait_500_ms` yields here, the
other branch (namely `read_message_from_queue`) gets cancelled, but when
we re-enter `read_message_from_queue`, we don't know where we left off.
That's a problem because we may have already read part of a message, and
upon re-entry, we want to continue reading from where we stopped.

To deal with this, the patch adds a new `AsyncReader` which is just a
stateful version of `async_read`. `AsyncReader` is cancel-safe because
it preserves state; when you call `AsyncReader::read`, it continues
based off the last state of the reader, rather than re-starting as
`async_read` does.

I've also audited all the opportunitic select branches. We haven't run
into this problem before because either
- the branches didn't use `async_read`/`async_write`, and only used
  cancel-safe methods
- the other branches took paths that would never re-enter into a
  cancelled call to `async_read`, so not being cancel-safe is okay.

Test:
- added a unit test to make sure the reader is cancel-safe
- added fuzz tests to fuzz for cancel safety (runs 1000 times per test run)
- `async_read` is now implemented in terms of `AsyncReader`

We don't need to do this for `async_write` because it's not currently
ever a part of this opportunistic branches. If that changes in the
future we should be cautious of it.

* Stray dbgs and lints