Clean up some todos and save some space (#432)
* Decrease memory size of `WorkId` and make it unique per test

Previously we weren't setting `WorkId` correctly on a per-manifest-entry
basis! Also, since WorkId is ABQ-internal, we don't need an expensive
repr like a String, 16 bytes is enough.

* Remove two needless fields from `WorkerTest`

These are holdovers from very very early days of ABQ and should have
been removed a long time ago. This will save a large amount of space per
test entry.

* Lints

* Throw away procspawn

* Remove obsolete todos

* Toss out `work_retries`

This is never used today and is a hold-over.

* Remove work_timeout

This is a holdover of early days ABQ and no longer needed

* Remove more stale todos