Handle test run cancellation before or during manifest reporting (#445)
* Handle test run cancellation before or during manifest reporting

Previously we assumed that when a manifest was received, a test run was
active. This is an incorrect assumption, as the test run may be
cancelled. This patch corrects that so that we handle cases where a test
run is cancelled before or during the time a manifest is received.

It also changes some locking and ordering conditions so that we handle
races on the ABQ queue state by taking exclusive access over the
supervisor responder(s) when we need to change the state of a run due to
a new manifest, or due to cancellation.

Closes #443

* bump ci

* bump ci