Implement `set-exit-code` (#460)
* Add support for abq set-exit-code

* Add integration tests for trying to set-exit-code after run completes

* Do not have in-band workers ask queue for exit code is supervisor is in-band

* Do not set exit code in-band if ABQ_SET_EXIT_CODE=false

* Handle potential race when a worker connects while OOB exit code is awaited

Prior to this patch there was a bug where if a worker connected while
the queue was waiting for an out-of-band exit code, we'd hit a marked
illegal state. But of course this is not an illegal state. This fix is
to have the worker poll until the exit code is known in this case.

* Add integration test for successful set-exit-code

* Remove unused import

* Fix compile error when matching RunWasCancelled.

---------

Co-authored-by: Doug Mayer <doug@rwx.com>