Simplify test-fetching interfaces and persisted worker connections (#465)
* Restructure test-fetching interfaces and persisted worker connections

Previously, we had multiple interfaces for fetching tests between the
generic test runner and ABQ worker levels. Additionally, the persistent
connections kept between workers and the queue for the purposes of
fetching new tests were functionalized and a bit confusing.

By defunctionalizing into structs, the implementation is clearer and
more efficient. We can now more explicitly represent that a persistent
connection from a worker to the queue must be owned uniquely, so there
is no need for atomic reference counts or mutexes. Hopefully, the new
implementation will be easier to read too.

* Lift persistent worker connections to their own module

* Address lints