Avoid unwrapping possible error when building request (#485)
`try_clone` a request from the reqwest library can fail - as far as I
can see, this can only happen when the build request itself has an
error. Nevertheless, there is no reason for us to suffer a potential
panic unwrapping the cloned request. We can instead build it fresh each
time, by passing a generator, which this patch does.