Log worker idle time after completion (#492)
* Always log run_id not wanted_run_id

* Inject notification dependency

* Add associative list container

This is much faster and uses less memory than a hash map when we need
it.

* Add log_assert macro

This performs runtime assertions but will log, rather than blowing up,
if an assertion fails.

* Log idle time between when workers complete and test run completes

Queue will keep timestamps of when workers complete in-memory until a
test run is seen as complete. When the test run completes, we pop the
timestamps and log the idle time for each worker recorded.

Note that some workers' timestamps may not be logged, when workers
notification of completion races with the test suite run completing.
This happens precisely when the worker end time is very close to the
test suite run completion time, which is not the pathological cases we
care about. For this reason, we don't lock on expecting all workers
reporting their completion time before we close up a test suite run.

* Perf: run non-racing async tasks concurrently

* Perf: run native test runner finalization concurrently

* Add tests for logged worker idle seconds