Ensure supervisor ticks do not interfere with timeout polls (#509)
* Ensure supervisor ticks do not interfere with timeout polls

Every so often the supervisor will "tick" the ABQ reporters; this is
done for reasons of interactive test output. However, the previously
implementation did so in a way that interefered with the inactivity
timeout. Rather than sleeping on the inactivity timeout interval, set a
moment in time when the inactivity timeout should fire, and guard on
that rather than branching on sleeping either the tick or inactitivy
timeout.

This patch could use an additional test that makes sure that after the
supervisor receives a message from the queue, the poll timeout is reset.

Closes #489

* assign reasonable next_poll_timeout_at immediately to avoid it ever holding an invalid value

Alternatively we could set it to `None` but it really should never be invalid

---------

Co-authored-by: Michael Glass <me@mike.is>