Persist manifests to s3 from disk
Rather than duplicating the manifest in memory for local and remote
persistence, have remote persistence read from the local path once it's
stored. This trades off less memory pressure for slower remote
persistence, which I belive is reasonable in this case, since manifests
are not likely to be needed immediately remotely.