Drop use of an LRU cache for results files
Using an LRU cache of file descriptors for results files could induce
the following race:

- A file P is opened, put in a mutex M, and placed into the cache. We hand to
  one thread 1 the mutex M.
- Thread 1 takes exclusive access of M.
- M is evicted from the cache. Thread 1 still has exclusive access to M.
- Thread 2 requests access to P from the cache. P is opened **in a new
  file descriptor** and placed in a new mutex O.
- Thread 2 takes exclusive access of O.
- Thread 1 and Thread 2 can now have unsynchronized writes to their file
  descriptors, which are different but point to the same filesystem
  inode.

This was a race to unlikely occur in practice, but it appears that it
can in fact occur.

The solution presented in this patch is to not maintain an LRU cache,
but a cache of file descriptors that never evicts locked file
descriptors. Once the results file for a run ID is entered into the
cache, the entry will never be removed. An entry can be in two states:

- Open: the file descriptor is open, and this is the unique file
  descriptor an ABQ queue has to a results file.
- Offloaded: the file was offloaded to a remote and no file descriptor
  is available until it is loaded from the remote.

This implementation trades off speed of access for memory usage. We
should expect that memory usage will grow steadily, and at a small,
consistent rate similar to that of the main queue state (that is, linear
to the number of run IDs on a queue). Since the cache entry is a string
and file descriptor + lock, I don't expect this to constitute a serious
source of memory leaks.

Alternative implementations to be considered:

- Use `flock` instead. This is likely too slow for our use case, since
  writes are frequent to result files when active. Also, this will not
  work on NFS.
- Perform all writes on a single thread and use channels to communicate
  messages. This could backup reads of test results significantly if
  there are many writes (including to other result files) ahead of it.
- Use a database (e.g. SQLite) instead