Drop support for ABQ statefiles
ABQ used to write a statefile to ABQ_STATE_FILE if the environment
variable was present. This was intended for integration with
[Captain](https://github.com/rwx-research/captain) in a legacy design of ABQ, and
was never removed.

Technically, this is a breaking change, but in practice we do not
anticipate that any customers of ABQ use this feature, as it was never
publicly advertised.