Correct flaky network test
`connecting_to_work_server_with_no_auth_fails` has a client send to the
queue a request without authentication, and then checks that the queue
observes a failed connection. However, the prior implementation did not
necessarily wait for the request to be received by the server, which
could cause flakiness.