Log warnings when workers have different commands
The command passed to a native test runner is now cryptographically hashed
and stored alongside a run ID. If a worker connects and has
a command different from the command given to the worker that
initialized run, log a warning on both the worker and the server.

This is backwards compatible. Older runs loaded on a queue won't have
their command hashes checked.