Make sure to communicate native runner info even if tests fail
Prior to this commit, an ABQ worker that observed a native runner
failure would bail with an error, without communicating the native
runner's information. This could pose a problem, because if the worker
ran any tests, we attempt to write to reporters regardless of whether
the worker succeeded or errored. As such, when it is available, make
sure to bubble up the native runner information to the reporters, even
in error scenarios.