Report hosted queue errors regardless of token kind
Prior to this patch, ABQ would fall back to attempting to using an
ephemeral queue if an organization access token was used with the hosted
API and the hosted API returned an unsupported queue error. However, ABQ
should exit unconditionally with the error message in such cases.

A test for `abq test` and `abq report` is included.