Ensure that runs with organization access tokens never use ephemeral queues
Some time ago, we added support for using personal access tokens with
ABQ runs. When a personal access token is used with an ABQ run, the
behavior is:

- If the run ID was determined by the CLI or inferred from an env var:
  use the response by the hosted API. If
  the hosted API instructs ABQ to run, do so in readonly mode; otherwise
  if the hosted API instruct ABQ that the mode is unsupported, exit.

- If the run ID was not explicitly set on the CLI or inferred from CI
  env vars, instead fall back to using an ephemeral ABQ instance for the
  run.

In a version of ABQ 1.6.4, the behavior was refined such that if the
hosted API returned an error, we would always display that error unless
a PAT without an explicit run ID was being used. Only in that case would
we fall back to the ephemeral queue mode.

However, we were missing test coverage to ensure that the use of
organization access tokens with implicit run IDs would use the hosted
queue, and so this introduced a breaking change. We did not have test
coverage for this, and RWX-internal usage of ABQ did not have any test
runs that use organization access tokens but auto-generated run IDs, so
this was not caught.

This PR corrects the behavior and introduces tests to handle those
cases. The logic for determining whether what we should do based on an
access token is re-written to have less branches and hopefully make it
much easier to follow.

- Org token, run ID from env: always use remote configuration
- Org token, run ID from CLI: always use remote configuration
- Org token, run ID auto-generated: always use remote configuration

- PAT, run ID from env: always use remote configuration
- PAT, run ID from CLI: always use remote configuration
- PAT, run ID auto-generated: always use ephemeral